Convert BranchDirEntry to tsx
